refactor: domain 위치 변경
