+++ backend/src/main/java/aiin/backend/domains/member/domain/Member.java
@@ -1,4 +1,4 @@
-package aiin.backend.domains.member;
+package aiin.backend.domains.member.domain;
 
 import aiin.backend.common.entity.BaseEntity;
 import jakarta.persistence.*;
